+++ tests/unit/test_imagehash.py
@@ -2,8 +2,9 @@
 __author__ = 'Lene Preuss <lene.preuss@gmail.com>'
 
 from pathlib import Path
-from typing import List
+from typing import List, Callable, Any
 
+import imagehash
 import pytest
 
 from duplicate_images.function_types import Results
@@ -16,7 +17,7 @@ def test_sequential(image_files: List[Path], algorithm: str) -> None:
     equals = ImagePairFinder.create(
         image_files, IMAGE_HASH_ALGORITHM[algorithm]
     ).get_equal_groups()
-    check_results(equals)
+    check_results(equals, algorithm)
 
 
 @pytest.mark.parametrize('algorithm', list(IMAGE_HASH_ALGORITHM.keys()))
@@ -25,7 +26,7 @@ def test_parallel(image_files: List[Path], algorithm: str) -> None:
         image_files, IMAGE_HASH_ALGORITHM[algorithm],
         options=PairFinderOptions(parallel=True)
     ).get_equal_groups()
-    check_results(equals)
+    check_results(equals, algorithm)
 
 
 @pytest.mark.parametrize('algorithm', list(IMAGE_HASH_ALGORITHM.keys()))
@@ -33,7 +34,7 @@ def test_max_distance(image_files: List[Path], algorithm: str) -> None:
     equals = ImagePairFinder.create(
         image_files, IMAGE_HASH_ALGORITHM[algorithm], options=PairFinderOptions(max_distance=1)
     ).get_equal_groups()
-    check_results(equals)
+    check_results(equals, algorithm)
 
 
 @pytest.mark.parametrize('algorithm', list(IMAGE_HASH_ALGORITHM.keys()))
@@ -42,7 +43,7 @@ def test_explicit_hash_size_works(image_files: List[Path], algorithm: str) -> No
         image_files, IMAGE_HASH_ALGORITHM[algorithm],
         options=PairFinderOptions(hash_size=8)
     ).get_equal_groups()
-    check_results(equals)
+    check_results(equals, algorithm)
 
 
 def test_bad_hash_size_whash(image_files: List[Path]) -> None:
@@ -58,7 +59,7 @@ def test_max_distance_parallel(image_files: List[Path], algorithm: str) -> None:
         image_files, IMAGE_HASH_ALGORITHM[algorithm],
         options=PairFinderOptions(parallel=True, max_distance=1)
     ).get_equal_groups()
-    check_results(equals)
+    check_results(equals, algorithm)
 
 
 @pytest.mark.parametrize('parallel', [False, True])
@@ -72,11 +73,15 @@ def test_create_with_all_parameters(
         image_files, IMAGE_HASH_ALGORITHM[algorithm],
         options=PairFinderOptions(max_distance=max_distance, hash_size=hash_size, parallel=parallel)
     ).get_equal_groups()
-    check_results(equals)
+    check_results(equals, algorithm)
 
 
-def check_results(equals: Results) -> None:
+def check_results(equals: Results, algorithm: str) -> None:
     assert any('jpeg_' in pair[0].name and 'half_' in pair[1].name for pair in equals)
     assert any('png_' in pair[0].name and 'half_' in pair[1].name for pair in equals)
     assert any('jpeg_' in pair[0].name and 'png_' in pair[1].name for pair in equals)
-    assert not any('jpeg_' in pair[0].name and 'subdir_' in pair[1].name for pair in equals)
+    if algorithm != 'crop_resistant':
+        assert not any('jpeg_' in pair[0].name and 'subdir_' in pair[1].name for pair in equals), [
+            ['/'.join(p[0].parts[3:]), '/'.join(p[1].parts[3:])]
+            for p in equals if 'subdir_' in p[1].name
+        ]