+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.11.5] - 2024-12-16
+
+### Added
+- crop-resistant hash algorithm with `--algorithm=crop_resistant`
+
+### Updated
+- Updated dependencies to fix security vulnerabilities
+
 ## [0.11.4] - 2024-12-16
 
 ### Updated